+++ build.gradle
@@ -42,11 +42,11 @@ test {
 }
 
 application {
-    mainClass.set("duke.gui.Launcher")
+    mainClass.set("cody.gui.Launcher")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "cody"
     archiveClassifier = null
 }
