+++ src/main/java/cody/gui/DialogBox.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package cody.gui;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -56,7 +56,7 @@ public class DialogBox extends HBox {
         return new DialogBox(text, img);
     }
 
-    public static DialogBox getDukeDialog(String text, Image img) {
+    public static DialogBox getCodyDialog(String text, Image img) {
         var db = new DialogBox(text, img);
         db.flip();
         return db;
