+++ src/main/java/cody/task/Deadline.java
@@ -1,10 +1,10 @@
-package duke.task;
+package cody.task;
 
 import java.time.LocalDate;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import duke.exception.CodyException;
+import cody.exception.CodyException;
 
 /**
  * Represents a deadline task.
